feat: support assembly implementation of euc (#2486)
* feat: support assembly implementation of euc

This puts in place the necessary support for the assembly implementation
of the EUC module.  Mostly, this comes down to remove the need to
compute certain intermediate columns which no longer exist.

* minor tweak to prevent oom in blockchain tests

The blockchain tests appear to run out-of-memory occasionaly (which is
separate from this PR); however, reducing the amount of parallelism
should resolve the issue.