update reference tests build.gradle
This updates the build.gradle to ensure that the
downloadExecutionSpecFixtures task is triggered when running the tests,
when previously it was not.  There is still an oddity though, whereby
the first time you do it, it will download the fixtures but then not run
any actual tests.  However, if you rerun then it does indeed run the
tests.