support in memory trace generation
This alters the way in which trace generation occurs, in that it traces
are always generated fully in memory before being written to disk.  This
allows various possibilities.  For example, we can compress the data
prior to writing to disk.  Or, we can use an alternative format which
requires holistic information of the entire trace before being written
to disk (e.g. it uses some kind of heap, etc).