change json encoding of blobBaseFees
This updates the json encoding of blobBaseFees to use a straightforward
array encoding (i.e. byte[]) which can be automatically managed by the
GSON decoder.  This prevents the need to develop a custom decoder for
Bytes, and prevents the replay files from being dependent upon the
internal representation of this class.